% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sj5_matchdata}
\alias{sj5_matchdata}
\title{sj5_matchdata}
\format{
A tibble with 744 rows and 10 columns.
\describe{
\item{id}{The ID of the match. This corresponds with the id field in metadata.}
\item{Character}{The name of the team the player was part of during the match.}
\item{Player}{The in-game name of the player.}
\item{Weapon}{The weapon the player used in the match.}
\item{Kills}{The number of splats the player secured in the match.}
\item{Assists}{The number of assists the player secured in the match.}
\item{Deaths}{The number of times the player was splatted during the match.}
\item{Specials}{The number of specials used by the player during the match.}
\item{Points}{The amount of turf the player painted during the match.}
}
}
\source{
\url(https://docs.google.com/spreadsheets/d/1i4IdNALs3fnsA4-2HSsyBQOrHCTBPNiSGzRzfnKdocE/edit?gid=1680411559#gid=1680411559)
}
\usage{
sj5_matchdata
}
\description{
A dataset consisting of player stats, taken from matches in Superjump 5.
}
\examples{
#'data(sj5_matchdata)
}
\keyword{datasets}
