% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{superjump5}
\alias{supeRjump5}
\alias{supeRjump5-package}
\alias{superjump5}
\alias{sj5}
\title{superjump5}
\description{
A package containing player and match information from
the alpha bracket of the SuperJump 5 Splatoon 3 major.
This dataset consists of two tibbles, sj5_matchdata and sj5_metadata, which are designed
to be used in conjunction with one another.
}
\details{
\describe {
\item{sj5_matchdata}{Player statistics from every match in Superjump 5's Alpha bracket.}
\item{sj5_metadata}{Information about each match in Superjump 5's Alpha bracket.}
}
}
\author{
\strong{Maintainer}: Spore \email{splatsporecloud@gmail.com}

}
